Fix #20 and correct treatment of compound flags
This resource had three major bugs that I sussed out while investigating and fixing this. 
1: Test string-compared existing rights as a single string against an array, which was the cause of the (almost)-always-fails-the-test bug. 
Repro was:
Set "Read" on a folder. The next 'Test' would retrieve the value 'Read, Synchronize', as synchronize is automatically added with Read by the OS.
@('Read') was then compare-object'd against @('Read, Synchronize'), causing the obvious test failure.
However the fix wasn't as easy as switching the compare object to @('Read, Synchronize') -split ', ' due to the fact that the Enum is flag based which leads us to #2. 

2: The script didn't properly take into account the subset-ness of the Flagged enum. 
For example, manually set permissions on a folder granting something FullControl, and then create a mof assigning "Read" to that same folder. The test fails, but the 'Set' does not correct the issue because applying Read to something with FullControl does nothing. -band'ing the values will reveal that 'Read' is fully contained as a subset of 'FullControl', so I corrected this as well. However this led me to #3...

3:  If the 'Requested' Rights are subsets of each other, the Test would also always fail. For example, assigning
@("Modify", "Write") to a folder with no permissions actually just results in @("Modify") in the resulting Enum, as "Write" is a fully contained subset of "Modify" and thus wont be returned individually in the stringified version of the enum. The fix for that was pre-converting the input "Rights" to the FileAccess enum and then diffing the enums directly. This was more or less already fixed by my original implementation of the 1 and 2 fixes, however required the special foreach loop handling in Ensure Absent to properly handle.

I'm not sure what the old behavior was on multi-permission ensure absent before (given that multi-permission didn't work before I'd consider that behavior undefined), but the behavior now is that if ANY of the (pre-enum-string) Ensure Absent permissions are wholly contained within the existing permissions, it will return false. The Set continues to remove the full ACL as it did before, regardless of what subset of existing permissions were requested to be removed by the Ensure Absent Rights.
So for a few examples: 
Test @("FullControl") Absent returns TRUE for a folder with existing @("Read", "Write") permissions, as FullControl is not fully contained by Read, Write. 
Test @("Read") Absent returns FALSE for a folder with existing @("Modify") permissions, as Read is fully contained by "Modify".
Test @("Read","Write") Absent returns FALSE for a folder with existing @("Modify") permissions, as either one is fully contained by "Modify".
Test @("Read","FullControl") Absent returns FALSE for a folder with existing @("Modify") permissions, as Read is fully contained by "Modify". # Note this is the reason for the Foreach loop in the Absent switch. Pre-combining "Read, FullControl" into the enum value would just yield "FullControl", which would cause the Test to pass as FullControl is not fully contained by "Modify".
Test @("FullControl") Absent returns TRUE for a folder with existing @("Modify") permissions, as FullControl is NOT fully contained by "Modify", even though they share "Read", "Write", etc.

I also improved the logging output, as I felt in the dark about what it was doing or how it was handling and combining the flags. Hopefully this is to your liking.